Texture-Manager - 과제
2차원 스프라이트 적용하기

+++ src/TextureManager.cpp
@@ -32,8 +32,8 @@ void TextureManager::draw(std::string id, int x, int y, int width, int height,
     SDL_RenderCopyEx(pRenderer, m_textureMap[id], &srcRect, &destRect, 0, 0, flip);
 }
 
-void TextureManager::drawFrame(std::string id, int x, int y, int width,
-    int height, int currentRow, int currentFrame,
+void TextureManager::drawFrame(std::string id, int x, int y, double width,
+    double height, int currentRow, int currentFrame,
     SDL_Renderer* pRenderer, SDL_RendererFlip flip)
 {
     SDL_Rect srcRect;
